Remove DOJ seal img alt text to skip
Ran WAVE testing on the website and it flagged the seal as having a long alt text. Indeed it has a long and beautifully detailed description with nuances I didn't realize myself, but it might not be appropriate because it's purely decorative and the name Civil Rights Division is beneath it. 
Reference: https://webaim.org/techniques/alttext/#decorative (Example 2)
If it were functional (ex: links to DOJ home page) we should add a label again.

How m' doin', Lindsay?