Refactor ShowView, add modified_date for reports
To leverage django's auto_now field attribute to
automatically update `modified_date` we need to use
the model's save method. ModelForms accomplish that
and are a common Django pattern for handling POST methods.

We now use the existing ComplaintActions form to handle
report updates as well as updating the activity stream