Skip 0049 record update migration if none exist
Adding the .exists() check here allows
us to skip the evaluation of the Report queryset
while running migrations on a new database.

Necessary because the data migration is not
aware that `report.modified_date` does not
yet exist and evaluated the update_records
queryset will fail when it attempts to pull
that field from the database.