Don't require auth for 404 responses
We're currently using django-auth-adfs's
`LoginRequiredMiddleware` to require login for
internal pages. As currently implemented, the middleware
doesn't readily allow for exempting 404 or other non-successful
responses from the authentication check.

Here we override the middleware to return a 404 response prior to
handling by the underlying django-auth-adfs Middleware.

Overriding this middleware isn't a documented path in djang-auth-adfs
so we'll pin our version to the current implementation to avoid unexpected
changes.