Return 422 error if form has been refreshed
If, while progressing through the form, a
user's session data has been cleared
- commonly caused by refreshing the page or
any other GET request - they'll need to
start the process from the begining.

By examining the management form contents
and comparing to current state of the data
we can detect occurences as soon as the first
POST following a refresh.