feat(judicial districts): add human-readable label text
A note on the implementation for this (https://github.com/usdoj-crt/crt-portal-management/issues/1098)

The select widget template never displayed label text, only option values.
This meant that simply updating DISTRICT_CHOICES to include readable label
text could not work without updating the template. It's possible this might
be a bug, so I went ahead and made that change.

This affected the section choices, which now displayed the full label. However,
we also do have a way of reducing section choices to acronym only, so we shift
to using that one in the dropdown.

Lastly, it seemed that the template title-cased its choices while individual
select widgets re-applied capitalization to counteract it. This also seemed
like emergent complexity that could be refactored. Select options now no longer
title-case its labels. With that change, option labels continue to display
as intended and we also no longer require a .text-uppercase helper class on
certain widgets.