refactor(settings): update django-compressor settings
Resolves the following deprecation warnings:

DeprecationWarning: COMPRESS_JS_FILTERS setting is deprecated and will be removed in django-compressor 3.0. Use COMPRESS_FILTERS instead.
DeprecationWarning: COMPRESS_CSS_FILTERS setting is deprecated and will be removed in django-compressor 3.0. Use COMPRESS_FILTERS instead.

We currently don't use django-compressor but this updates the syntax
in case we ever want to use it