Update existing templates to use referral_text
- 🌎 The referral templates currently use a hard-coded referral address
- ⛔ This isn't great because these addresses and numbers change often, and we'd
  like to be able to update them from the admin panel.
- ✅ This commit moves those addresses to referral_text variables on the
  relevant ReferralContact objects

This was done by searching for "referral_contact: va", etc, and then
running a macro to grab the text for the  given translation, drop
it into the migration, and replace its spot in the template with the
variable. This was somewhat manual (I had to choose the correct language), but mostly automated.